requirements.txt
aiohttp==3.7.4.post0
astroid==2.11.5
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
cohere==2.5.0
dill==0.3.5.1
google-search-results==2.4.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
multidict==6.0.2
numpy==1.22.4
platformdirs==2.5.2
pylint==2.13.9
PyYAML==6.0
requests==2.27.1
soupsieve==2.3.2.post1
tomli==2.0.1
typing_extensions==4.2.0
streamlit==1.12.0
streamlit-chat==0.0.2.1
urllib3==1.26.9
wrapt==1.14.1
yarl==1.7.2
evaluate==0.2.2
discord.py==2.0.1